+++ changelog.md
@@ -53,6 +53,7 @@
                   throw new IllegalArgumentException("expected");
               }));
   ```
+* Added a new `execute` Kotlin extension function to the `ApplicationService` that allows one to use a `java.util.UUID` as a streamId when working with lists of events (as opposed to `Sequence`).   
   
 
 ### 0.16.3 (2023-05-12)