+++ changelog.md
@@ -53,7 +53,8 @@
                   throw new IllegalArgumentException("expected");
               }));
   ```
-* Added a new `execute` Kotlin extension function to the `ApplicationService` that allows one to use a `java.util.UUID` as a streamId when working with lists of events (as opposed to `Sequence`).   
+* Added a new `execute` Kotlin extension function to the `ApplicationService` that allows one to use a `java.util.UUID` as a streamId when working with lists of events (as opposed to `Sequence`).
+* Upgraded xstream from 1.4.19 to 1.4.20
   
 
 ### 0.16.3 (2023-05-12)