+++ changelog.md
@@ -55,8 +55,8 @@
   ```
 * Added a new `execute` Kotlin extension function to the `ApplicationService` that allows one to use a `java.util.UUID` as a streamId when working with lists of events (as opposed to `Sequence`).
 * Upgraded xstream from 1.4.19 to 1.4.20
-* Added better logging to `org.occurrent.subscription.mongodb.spring.blocking.SpringMongoSubscriptionModel`, including some debug logs that can be used for detailed information about what's going on.
-  
+* Added better logging to `org.occurrent.subscription.mongodb.spring.blocking.SpringMongoSubscriptionModel`, including some debug logs that can be used for detailed information about what's going on. 
+* Upgraded Kotlin from 1.8.21 to 1.9.0
 
 ### 0.16.3 (2023-05-12)
 * Added support to the retry module to execute retries with a function that takes an instance of `org.occurrent.retry.RetryInfo`. This is useful if you need to know the current state of your of the retry while retrying. For example: