+++ changelog.md
@@ -58,6 +58,7 @@
 * Added better logging to `org.occurrent.subscription.mongodb.spring.blocking.SpringMongoSubscriptionModel`, including some debug logs that can be used for detailed information about what's going on. 
 * Upgraded Kotlin from 1.8.21 to 1.9.0
 * Upgraded Spring Boot from 3.0.6 to 3.0.8
+* Upgraded Spring Aspects from 6.0.9 to 6.0.10
 
 ### 0.16.3 (2023-05-12)
 * Added support to the retry module to execute retries with a function that takes an instance of `org.occurrent.retry.RetryInfo`. This is useful if you need to know the current state of your of the retry while retrying. For example: